testing/test-requirements.txt
robotframework
robotframework-requests
Jinja2
requests
urllib3
lxml
charset-normalizer
certifi
idna
MarkupSafe
